complete a invoices page and create a team page.
